A gcc warning

+++ src/RegexSearch.h
@@ -44,11 +44,11 @@ public:
     Match(){m_start=wxNOT_FOUND;m_length = 0;}
     void SetStart(size_t start){m_start = start;}
     void SetLength(size_t length){m_length = length;}
-    size_t GetStart() const {return m_start;}
+    long long GetStart() const {return m_start;}
     size_t GetLength() const {return m_length;}
     size_t GetEnd() const {return m_start + m_length;}
   private:
-    size_t m_start;
+    long long m_start;
     size_t m_length;
   };
   explicit RegexSearch(wxString regex);
